requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.10
langgraph==0.0.20
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.5.0
python-multipart==0.0.6 